requirements.txt
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.7.2
asn1crypto==1.5.1
binaryornot==0.4.4
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
cryptography==42.0.5
cssselect2==0.7.0
Django==5.0.2
django-browser-reload==1.12.1
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-tailwind==3.8.0
django-truncate==0.1
html5lib==1.1
idna==3.6
Jinja2==3.1.3
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
oscrypto==1.3.0
pdfkit==0.6.0
pillow==10.2.0
pycparser==2.21
Pygments==2.17.2
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
pypdf==4.0.2
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-slugify==8.0.4
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
rich==13.7.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.2.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.1
uuid==1.30
webencodings==0.5.1
xhtml2pdf==0.2.15
